Move parts of mila init command body to new module
